requirements.txt
streamlit>=1.20.0
google-genai
